refactor: implement `TryFrom` trait for Property->ParamType conversion (#469)
* Move parsing helpers in their own module for better separation. 
* Implement `TryFrom` trait for Property->ParamType conversion. This allows for cleaner and more idiomatic Rust code. 
* Rename the function selector hashing function, old name was confusing, it is actually just a util function.
